refactor(store): move getUrlData to a separate file
