+++ src/types/window.d.ts
@@ -30,4 +30,11 @@ interface Window {
     Ya?: {
         Rum?: RumCounter;
     };
+
+    // eslint-disable-next-line
+    web_version?: boolean;
+    // eslint-disable-next-line
+    custom_backend?: string;
+
+    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof import('redux').compose;
 }
